Maze Solver Completed Project
Here are 2 Android applications.
The one in maze_writer generates a text file with a randomized maze and
places it in the Documents directory.
The one in src not only reads and builds a maze from the text files it
found, but also prints it and solves the maze using both blind and
heuristic algorithms.